+++ App.js
@@ -5,10 +5,10 @@ import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
 import Ionicons from 'react-native-vector-icons/Ionicons';
 import * as eva from '@eva-design/eva';
 import { ApplicationProvider } from '@ui-kitten/components';
-import ChoresScreen from './components/ChoresScreen/ChoresScreen';
-import CalenderScreen from './components/CalenderScreen/CalenderScreen';
-import AlbumScreen from './components/AlbumScreen/AlbumScreen';
-import MyHomeScreen from './components/MyHomeScreen/MyHomeScreen';
+import ChoresScreen from './components/ChoresScreen';
+import CalenderScreen from './components/CalenderScreen';
+import AlbumScreen from './components/AlbumScreen';
+import MyHomeScreen from './components/MyHomeScreen';
 
 const Tab = createBottomTabNavigator();
