+++ components/AlbumScreen.js
