+++ components/CalenderScreen.js
@@ -2,9 +2,9 @@ import React, { useState } from 'react';
 import { SafeAreaView } from 'react-native';
 import { Button, Layout, Text } from '@ui-kitten/components';
 import SubviewSwitchBar from './SubviewSwitchBar';
-import MonthlyCalenderView from './MonthlyCalenderView/MonthlyCalenderView';
-import WeeklyCalenderView from './WeeklyCalenderView/WeeklyCalenderView';
-import DailyCalenderView from './DailyCalenderView/DailyCalenderView';
+import MonthlyCalenderView from './MonthlyCalenderView';
+import WeeklyCalenderView from './WeeklyCalenderView';
+import DailyCalenderView from './DailyCalenderView';
 
 export default function CalenderScreen() {
   const [subviewIdx, setSubviewIdx] = useState(2);
