+++ components/ChoresScreen.js
