+++ components/DailyCalenderView.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { List } from '@ui-kitten/components';
 import DaySchedule from './DaySchedule';
-import { eventsData } from '../../../staticData';
+import { eventsData } from '../staticData';
 
 export default function DailyCalenderView() {
   return (
